modernbert: add ModernBERT architecture support
- Add ModernBERT architecture support for embedding models
- Implement alternating local/global attention patterns with sliding window attention
- Support dual RoPE theta values (local: 10000, global: 80000)
- Add GeGLU (gated FFN) support with proper tensor splitting
- Configure CLS pooling and GPT-2 tokenizer for sentence-transformers compatibility
- Handle 8192 context length with proper truncation support
- Fix tensor loading for output_norm and layer_out_norm tensors

+++ llama/llama.cpp/src/llama-graph.cpp
@@ -470,7 +470,7 @@ void llm_graph_input_attn_cross::set_input(const llama_ubatch * ubatch) {
             }
         }
 
-        for (int i = n_tokens; i < n_tokens; ++i) {
+        for (uint32_t i = n_tokens; i < (uint32_t)n_tokens; ++i) {
             for (int j = 0; j < n_enc; ++j) {
                 data[h*(n_enc*n_tokens) + i*n_enc + j] = -INFINITY;
             }
